+++ src/test/java/de/jfachwert/zeit/ZeitdauerTest.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright (c) 2023 by Oli B.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express orimplied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * (c)reated 10.07.23 by oboehm
+ */
+package de.jfachwert.zeit;
+
+import de.jfachwert.FachwertTest;
+import org.junit.jupiter.api.Test;
+
+import java.util.concurrent.TimeUnit;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.containsString;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+/**
+ * Unit-Tests fuer {@link Zeitdauer}-Klasse.
+ *
+ * @author oboehm
+ */
+public final class ZeitdauerTest extends FachwertTest {
+
+    @Test
+    @Override
+    public void testToString() {
+        Zeitdauer zeitdauer = new Zeitdauer(5, TimeUnit.DAYS);
+        String s = zeitdauer.toString();
+        assertEquals("5 Tage", s);
+    }
+
+    @Test
+    void testToStringAllUnits() {
+        for (TimeUnit unit : TimeUnit.values()) {
+            Zeitdauer zeitdauer = new Zeitdauer(10, unit);
+            String s = zeitdauer.toString();
+            assertThat(s, containsString("10"));
+        }
+    }
+
+}